fix(engines): the write path's leak shape, and three tests that could not fail
Review of the commit below. The headline is a leak shape in the wrapper that
commit added, and three tests that asserted hazards they could not detect.

THE LEAK SHAPE. `ObjectHandle::replace_stream_data` took `&Document` beside
`&self` so it could reach the error slot. `ObjectHandle` carries no document
identity, so nothing stopped a caller passing a DIFFERENT document -- qpdf
latches on `self.data`'s slot, the wrapper drains the other document's empty
one, and the function returns `Ok(())` on a write that did not happen. That
is the sentence this milestone exists to prevent. The slot is now reached
through `self.data`, and `filter`/`decode_parms` are checked against it:
qpdf handles are bare `++next_oh` counters, so a handle from another document
would likely COLLIDE with a live one here and be written as `/Filter` with no
error at all.

THREE TESTS THAT COULD NOT FAIL, each found by planting the mutation it
exists for:

- the `_malloc(0)` guard: the fake's allocator did `Math.max(n, 1)`, so it
  could never return 0 and deleting the guard stayed green. It returns 0 for
  a zero request now, which is the behaviour C permits and the guard defends.
- the use-after-free: `_free` only removed a set entry, and the bump
  allocator never reuses, so moving the free ABOVE the engine call -- the
  textbook defect here -- passed everything. It poisons the region now.
- the heap-growth case swapped in a view over a new live buffer, which is
  replacement, not detachment; its comment claimed a hoisted view "would
  throw" and nothing threw. `ArrayBuffer.transfer` detaches for real, and the
  hoist now fails with "set on a detached ArrayBuffer".

All five mutations from the review's sweep are caught; two of them survived
before.

CLAIMS CORRECTED, each measured false rather than reworded:

- `bridge_qpdf.rs` said wasm-bindgen makes "two copies" of the slice. It
  makes none -- `getArrayU8FromWasm0` is a `subarray` view -- and the comment
  contradicted `bridge.rs` in the same commit. It matters: if the slice were
  already copied, the read-HEAPU8-after-the-malloc rule reads as
  belt-and-braces rather than as the requirement it is.
- "`bridge-write-path.test.ts` asserts each one" of the five hazards. Three
  are asserted, one cannot be asserted from a JS test at all, and one is
  asserted on the Rust side. The rustdoc says which, now.
- the commit below said `burrow_wasm_bg.wasm` was unchanged because LTO
  strips code nothing calls. The symbol is in the built artifact --
  wasm-bindgen emits an import shim regardless -- and the module grew 16
  brotli. Sixteen is nothing rounded, not nothing measured.

ALSO: `Err` from either wrapper does not mean the stream was not written --
`qpdf_oh_replace_stream_data` is three nested trapped calls and an inner
failure latches on the same slot, so #131 must not retry on the assumption
the document is untouched. A filter handle qpdf does not recognise silently
becomes "no filter", which is invisible today because burrow only passes
null. The `qpdf_oh_new_null` exemption now states where it stops, including
`QTC::TC_real`'s `safe_fopen` clause. `buf` is u32-coerced per this bundle's
own convention. The test rig uses `runInNewContext` with `self === globalThis`
like `test-scope.ts`, so bare-identifier resolution is modelled, and reads
the file list from `stage-web-engines.mjs` instead of hardcoding it.

`size-budget.json` re-records every line that moved -- +696 worker, +195
qpdf.wasm, +16 wasm, -3 and -8 on the two page lines, +896 total -- all
inside the 2% tolerance and re-recorded anyway, because that tolerance is for
toolchain drift and this is content somebody wrote. NOT RAISED: 509,812
leaves 40,069 bytes, 7.9%.

+++ apps/web/src/worker/bridge-qpdf.js
@@ -196,7 +196,7 @@ self.__burrow_qpdf_oh_new_null = (data) => u32(qpdf()._qpdf_oh_new_null(data));
  */
 self.__burrow_qpdf_oh_replace_stream_data = (data, stream, bytes, filter, decodeParms) => {
   const module = qpdf();
-  const buf = module._malloc(bytes.length === 0 ? 1 : bytes.length);
+  const buf = u32(module._malloc(bytes.length === 0 ? 1 : bytes.length));
   if (buf === 0) {
     return false;
   }
